Allow reading of already queued lines even after pipe has closed (#63)
* Allow reading of already queued lines even after pipe has closed

Even if pipe is closed and it sets state to error
the readline should read all the already received and queued lines.
Otherwise in some cases last traces / output from process
is lost.

* Queue data from pipe even if it just closed.

This was made possible by calling NonBlockingStreamReader._read_fd() in
the else branch where the pipe has closed. The pipe must also be read
empty as this function will not be called anymore.